\name{acss2}
\alias{acss2}
\title{ACSS complexity}
\usage{
acss2(string)
}
\arguments{
  \item{string}{\code{character} vector containing the to
  be analyzed strings (can contain multiple strings).}
}
\description{
Functions to conveniently compute algorithmic complexity
for short string, an approximation of the Kolmogorov
Complexity of a short string using the coding theorem
method.
}

