\docType{data}
\name{acss_data}
\alias{acss_data}
\title{acss: algorithmic complexity of short strings}
\usage{
acss_data
}
\description{
Contains the algorithmic complexity for short string, an
approximation of the Kolmogorov Complexity of a short
string using the coding theorem method.

This dataset shouldn'T be called directly but rather
through the acessor functions such as \code{\link{acss}}.
}
\author{
Nicolas Gauvrit, Fernando Soler Toscano, and Hector Zenil.
Ported to R by Henrik Singmann.
}

