% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm.R
\name{bdm}
\alias{bdm}
\title{Block Decomposition Method}
\usage{
bdm(string, delta = 1, alphabet = 9, span = 4)
}
\arguments{
\item{string}{\code{character} vector containing the to be analyzed strings (can contain multiple strings).}

\item{delta}{distance between two windows. Default is 1.}

\item{alphabet}{\code{numeric}, the number of possible symbols (not necessarily actually appearing in str). Must be one of \code{c(2, 4, 5, 6, 9)} (must be a scalar value in the current implementation). Default is 9.}

\item{span}{size of substrings/windows to be created from \code{string}. Default is 4.}
}
\description{
Obtain Complexity of Longer Strings than allowed by ACSS via the Block Decomposition Method.
}
\examples{

acss("SSOOXFXFOXXOXOXFFXXXSOTTO")  # too long

bdm("SSOOXFXFOXXOXOXFFXXXSOTTO") # default span is 4
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 3)
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 6)
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 11)

multi <- c(
  "SSOOXFXFOXXOXOXFFXXXSOTTO",
  "SSODXFXDOXXOXOXFFXRRSORTO",
  "DXXXXRRXXXSSOOOXOFFFOOOOO"
)

bdm(multi)
bdm(multi, delta = 4, span = 5) 
bdm(multi, span = 6)

}

