% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm.R
\name{bdm}
\alias{bdm}
\title{Block Decomposition Method}
\usage{
bdm(string, delta = 1, alphabet = 9, span = 10)
}
\arguments{
\item{string}{\code{character} vector containing the to be analyzed strings (can contain multiple strings).}

\item{delta}{distance between two windows. Default is 1 corresponding to a full overlap of the substrings. If \code{delta == span} (the maximum) the decomposition is disjoint, hence a partition.}

\item{alphabet}{\code{numeric}, the number of possible symbols (not necessarily actually appearing in \code{string}). Must be one of \code{c(2, 4, 5, 6, 9)} (must be a scalar value in the current implementation). Default is 9.}

\item{span}{size of substrings/windows to be created from \code{string}. Default is 10.}
}
\description{
Obtain Complexity of Longer Strings than allowed by ACSS via the Block Decomposition Method.
}
\examples{

acss("SSOOXFXFOXXOXOXFFXXXSOTTO")  # too long

bdm("SSOOXFXFOXXOXOXFFXXXSOTTO") # default span is 10
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 4)
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 6)
bdm("SSOOXFXFOXXOXOXFFXXXSOTTO", span = 11) # gives warning

multi <- c(
  "SSOOXFXFOXXOXOXFFXXXSOTTO",
  "SSODXFXDOXXOXOXFFXRRSORTO",
  "DXXXXRRXXXSSOOOXOFFFOOOOO"
)

bdm(multi)
bdm(multi, span = 5, delta = 2) 
bdm(multi, delta = 5)


# binary bdm should give 57.5664 in this case 
bdm("010101010101010101", alphabet = 2, span = 12, delta = 1)

# uses a substring of size less than 12:
bdm("010101010101010101", alphabet = 2, span = 12, delta = 12)

}

